Initial changes for implementing annotation status

+++ backend/projects/views.py
@@ -364,12 +364,12 @@ class ProjectViewSet(viewsets.ModelViewSet):
         project_response.data["last_pull_result"] = last_project_export_result
 
         # Add a field to specify the no. of available tasks to be assigned
-        project_response.data["unassigned_task_count"] = get_task_count(pk, UNLABELED)
+        project_response.data["unassigned_task_count"] = get_task_count(pk, INCOMPLETE)
 
         # Add a field to specify the no. of labeled tasks
         project_response.data["labeled_task_count"] = (
             Task.objects.filter(project_id=pk)
-            .filter(task_status=LABELED)
+            .filter(task_status=COMPLETED)
             .filter(review_user__isnull=True)
             .exclude(annotation_users=request.user.id)
             .count()
@@ -433,16 +433,18 @@ class ProjectViewSet(viewsets.ModelViewSet):
 
         tasks = Task.objects.filter(
             Q(project_id=project.id) & Q(annotation_users__in=[user])
-        ).filter(Q(task_status="unlabeled") | Q(task_status="draft"))
+        ).filter(Q(task_status="incomplete"))
 
-        Annotation_model.objects.filter(
-            Q(completed_by=user) & Q(task__task_status="draft")
-        ).delete()  # delete all draft annotations by the user
+        if tasks:
+            Annotation_model.objects.filter(
+                Q(completed_by=user) & Q(task__in=tasks)
+            ).delete()  # delete all annotations for incomplete tasks by the user
 
-        for task in tasks:
-            task.annotation_users.remove(user)
+            for task in tasks:
+                task.annotation_users.remove(user)
+                task.save()
 
-        tasks.update(task_status="unlabeled")  # unassign user from tasks
+        #tasks.update(task_status="unlabeled")  # unassign user from tasks
 
         project.frozen_users.add(user)
 
@@ -474,9 +476,12 @@ class ProjectViewSet(viewsets.ModelViewSet):
             .filter(review_user=user)
             .exclude(task_status__in=[ACCEPTED, TO_BE_REVISED])
         )
-        for task in tasks:
-            task.review_user = None
-            task.save()
+
+        if tasks:
+            Annotation_model.objects.filter(Q(completed_by=user) & Q(task__in=tasks)).delete()
+            for task in tasks:
+                task.review_user = None
+                task.save()
 
         project.frozen_users.add(user)
         project.save()
@@ -585,13 +590,13 @@ class ProjectViewSet(viewsets.ModelViewSet):
                     Task.objects.filter(
                         project_id__exact=project.id,
                         review_user=request.user.id,
-                        task_status__exact=LABELED,
+                        task_status__exact=COMPLETE,
                     )
                     if is_review_mode
                     else Task.objects.filter(
                         project_id__exact=project.id,
                         annotation_users=request.user.id,
-                        task_status__exact=UNLABELED,
+                        task_status__exact=INCOMPLETE,
                     )
                 )
             else:
@@ -600,12 +605,12 @@ class ProjectViewSet(viewsets.ModelViewSet):
                 unattended_tasks = (
                     Task.objects.filter(
                         project_id__exact=project.id,
-                        task_status__exact=LABELED,
+                        task_status__exact=COMPLETE,
                     )
                     if is_review_mode
                     else Task.objects.filter(
                         project_id__exact=project.id,
-                        task_status__exact=UNLABELED,
+                        task_status__exact=INCOMPLETE,
                     )
                 )
 
@@ -794,24 +799,20 @@ class ProjectViewSet(viewsets.ModelViewSet):
             )
 
         # check if user has pending tasks
-        # the below logic will work only for required_annotators_per_task=1
-        # TO-DO Modify and use the commented logic to cover all cases
-        pending_tasks = (
-            Task.objects.filter(project_id=pk)
-            .filter(annotation_users=cur_user.id)
-            .filter(task_status__exact=UNLABELED)
-            .count()
-        )
+        pending_tasks = Task.objects.filter(project_id=pk).filter(annotation_users=cur_user.id).filter(task_status__exact=INCOMPLETE)
+
+        pending_tasks_count = Annotation_model.objects.filter(task__in=pending_tasks).filter(completed_by__exact=cur_user.id).filter(annotation_status=UNLABELED).count()
+
         # assigned_tasks_queryset = Task.objects.filter(project_id=pk).filter(annotation_users=cur_user.id)
         # assigned_tasks = assigned_tasks_queryset.count()
         # completed_tasks = Annotation_model.objects.filter(task__in=assigned_tasks_queryset).filter(completed_by__exact=cur_user.id).count()
         # pending_tasks = assigned_tasks - completed_tasks
-        if pending_tasks >= project.max_pending_tasks_per_user:
+        if pending_tasks_count >= project.max_pending_tasks_per_user:
             return Response(
                 {"message": "Your pending task count is too high"},
                 status=status.HTTP_403_FORBIDDEN,
             )
-        tasks_to_be_assigned = project.max_pending_tasks_per_user - pending_tasks
+        tasks_to_be_assigned = project.max_pending_tasks_per_user - pending_tasks_count
 
         if "num_tasks" in dict(request.data):
             task_pull_count = request.data["num_tasks"]
@@ -836,7 +837,7 @@ class ProjectViewSet(viewsets.ModelViewSet):
         tasks = Task.objects.filter(project_id=pk)
         tasks = tasks.order_by("id")
         tasks = (
-            tasks.filter(task_status=UNLABELED)
+            tasks.filter(task_status=INCOMPLETE)
             .exclude(annotation_users=cur_user.id)
             .annotate(annotator_count=Count("annotation_users"))
         )
@@ -855,6 +856,12 @@ class ProjectViewSet(viewsets.ModelViewSet):
         for task in tasks:
             task.annotation_users.add(cur_user)
             task.save()
+            base_annotation_obj = Annotation_model(
+                result=[],
+                task=task,
+                completed_by=cur_user,
+            )
+            base_annotation_obj.save()
 
         project.release_lock(ANNOTATION_LOCK)
         return Response(
@@ -878,11 +885,12 @@ class ProjectViewSet(viewsets.ModelViewSet):
                 tasks = (
                     Task.objects.filter(project_id__exact=project_id)
                     .filter(annotation_users=user.id)
-                    .filter(task_status=UNLABELED)
+                    .filter(task_status=INCOMPLETE)
                 )
                 if tasks.count() > 0:
                     for task in tasks:
                         task.unassign(user_obj)
+                    Annotation_model.objects.filter(completed_by=user).filter(task__in=tasks).delete()
                     return Response(
                         {"message": "Tasks unassigned"}, status=status.HTTP_200_OK
                     )
@@ -993,7 +1001,7 @@ class ProjectViewSet(viewsets.ModelViewSet):
         # check if the project contains eligible tasks to pull
         tasks = (
             Task.objects.filter(project_id=pk)
-            .filter(task_status=LABELED)
+            .filter(task_status=COMPLETE)
             .filter(review_user__isnull=True)
             .exclude(annotation_users=cur_user.id)
         )
@@ -1013,6 +1021,12 @@ class ProjectViewSet(viewsets.ModelViewSet):
         for task in tasks:
             task.review_user = cur_user
             task.save()
+            base_annotation_obj = Annotation_model(
+                result=[],
+                task=task,
+                completed_by=cur_user,
+            )
+            base_annotation_obj.save()
 
         project.release_lock(REVIEW_LOCK)
         return Response(
@@ -1037,11 +1051,12 @@ class ProjectViewSet(viewsets.ModelViewSet):
             if project_obj and user in project_obj.annotation_reviewers.all():
                 tasks = (
                     Task.objects.filter(project_id__exact=project_id)
-                    .filter(task_status=LABELED)
+                    .filter(task_status=COMPLETE)
                     .filter(review_user=user.id)
                 )
                 if tasks.count() > 0:
                     tasks.update(review_user=None)
+                    Annotation_model.objects.filter(completed_by=user).filter(task__in=tasks).delete()
                     return Response(
                         {"message": "Tasks unassigned"}, status=status.HTTP_200_OK
                     )
@@ -1235,7 +1250,7 @@ class ProjectViewSet(viewsets.ModelViewSet):
 
                 # get labeled task count
                 labeled_tasks = get_annotated_tasks(
-                    pk, each_user, "labeled", start_date, end_date
+                    pk, each_user, "complete", start_date, end_date
                 )
                 items.append(("Labeled Tasks", labeled_tasks.count()))
 
@@ -1246,7 +1261,7 @@ class ProjectViewSet(viewsets.ModelViewSet):
                 items.append(("To Be Revised Tasks", to_be_revised_tasks.count()))
 
             # get unlabeled count
-            total_unlabeled_tasks_count = get_tasks_count(pk, each_user, "unlabeled")
+            total_unlabeled_tasks_count = get_tasks_count(pk, each_user, "incomplete")
             items.append(("Unlabeled Tasks", total_unlabeled_tasks_count))
 
             # get skipped tasks count
@@ -1425,7 +1440,7 @@ class ProjectViewSet(viewsets.ModelViewSet):
             tasks = Task.objects.filter(project_id=project.id).filter(
                 task_status=ACCEPTED
             )
-            tasks.update(task_status=LABELED)
+            tasks.update(task_status=COMPLETE)
             project.enable_task_reviews = True
             project.save()
             return Response(
@@ -1464,7 +1479,7 @@ class ProjectViewSet(viewsets.ModelViewSet):
                 parent_annotation__isnull=True
             ).delete()
             # mark all unreviewed tasks accepted
-            unreviewed_tasks = tasks.filter(task_status=LABELED)
+            unreviewed_tasks = tasks.filter(task_status=COMPLETE)
             unreviewed_tasks.update(task_status=ACCEPTED)
             # unassign reviewers
             tasks.update(review_user=None)
